Move createViewState to its own file
