+++ packages/saved-views-react/src/experimental.ts
@@ -6,7 +6,6 @@ export { createSavedViewOptions, type CreateSavedViewOptionsParams } from "./Sav
 export { SavedViewsExpandableBlockWidget } from "./SavedViewsWidget/SavedViewsExpandableBlockWidget.js";
 export { SavedViewsFolderWidget } from "./SavedViewsWidget/SavedViewsFolderWidget.js";
 export type { LegacySavedViewBase } from "./api/utilities/SavedViewTypes.js";
-export {
-  applyExtensionsToViewport, createViewState,
-} from "./api/utilities/translation/SavedViewTranslation.js";
+export { applyExtensionsToViewport } from "./api/utilities/translation/SavedViewTranslation.js";
+export { createViewState } from "./createViewState.js";
 export { ModelCategoryOverrideProvider } from "./ui/viewlist/ModelCategoryOverrideProvider.js";
